+++ index.php
@@ -1,2 +1,20 @@
+<?php get_header(); ?>
+
+<div class="container mx-auto my-8">
+
+	<?php if ( have_posts() ) : ?>
+		<?php
+		while ( have_posts() ) :
+			the_post();
+			?>
+
+			<?php get_template_part( 'template-parts/content', get_post_format() ); ?>
+
+		<?php endwhile; ?>
+
+	<?php endif; ?>
+
+</div>
+
 <?php
-/** Silence is golden. */
+get_footer();