+++ postcss.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    plugins: [
+        require('postcss-nested-ancestors'),
+        require('postcss-nested'),
+        require('postcss-import'),
+        require('tailwindcss')
+    ]
+}