+++ resources/js/app.js
@@ -6,4 +6,3 @@ window.addEventListener('load', function () {
             main_navigation.classList.toggle('hidden');
       });
 });
-console.log('test123');